Use Stripe Payment Intents with stripe gem
This implementation is based on online guides found on Stripe website
showing how to integrate the Payment Intents payment system. It is
based on the gem `stripe`, as the guides were using it.
It would be optimal to be able to use `active-merchant` also for Payment
Intents, so we can remove the extra gem and hopefully simplify the model
code.

This is still WIP, but at this point the feature is basically working,
although there's some work to do:

* have a proper Stripe gem configuration, or remove that gem and use
  ActiveMerchant Stripe library, like V2 and V3 Elements do
* use same HTML as Stripe Elements version
* improve error handling
* beta testing