Add ShippingRatesService
This service object retrieves the available shipping methods
to be passed to Stripe frontend payment request that will be
shown to the customer.

We need to select all the shipping methods available for all
the order shipments and sum their rates, so the customer can
decide what service and what price suits their needs.