Add PaymentRequestController
This controller manages requests from the frontend payment request
when the cart page payment is enabled.

It provides the shipping methods and their cost to be shown to the
customer and an endpoint that process the order with the customer
and payment information received from the FE payment request, after
the customer selects their address and the shipping method for the
order.