Introduce card pre-authorization for Payment Intents
At the moment payments done with Payment Intents end up in a double
charge for the customer. A first charge is created istantly when
the customer authorizes their card, and a second one is created later
after order confirmation. This issue is not of easy solution, as the
first charge is always required for Payment Intents, and the second
one is created automatically by Solidus flow.

A possible temporary workaround for this issue is to introduce the
concept of card pre-authorization: the first charge, now limited to
only 1$ is not captured anymore, and have an appropriate description
so, even when customers receive notifications for uncaptured
transactions from their bank, they will understand this is not going
to be actually captured.