Replace test 3D-Secure card number
The card ending with 3184 is not compatible with real-life scenarions
as it will always require customer manual authorization, regardless of
the actual settings for the payment and payment method used.

So, the new card ending with 3155 simulates more realisticly what
happens in real life, being still 3D-Secure, but allowing payment
methods to create off-session charges.