Manual capture and off-session usage for Payment Intents
When the customer pays with Payment Intents we need to explicitly
require manual capture, or the Payment Intents charge will be
marked as authorized as soon as the customer authorizes the use
of their card.

Also, in order to allow the reusability of the payment method that
is created when the customer adds a new card, we need to explicitly
mark the payment intents valid for off-session usage.