Decorate Spree::OrderUpdateAttributes
We need to make sure that when the customer is checking out by
adding a new Payment Intents card, we don't create a second
payment by following the standard Solidus payment flow.

Also, when the customer does completes the payment by changing
the payment method after having added a new Payment Intents card,
we need to make sure that the pending charge for that intent is
voided.