Manage cart checkout in CreateIntentsPaymentService
A few changes are required in order to properly manage the
cart checkout flow in this service object.

Cart checkout page sends the address attributes as a JSON
that we need to manipulate before Solidus can use it to
create an address record. Also, the customer full name
needs to be extracted differently when checking out from
the cart page.