Manage payment failure creation in `create_payment`
When for any reason the Solidus payment cannot be created, we need
to notify the JS code that something went wrong, so the error message
can be shown to the customer.

When the payment cannot be created, we also need to cancel the payment
intent on Stripe, or it will remain pending and this will result in
blocking some funds on the customer's card.