Add StripeCreditCard#try_void
This method is now the preferred way in Solidus for canceling a payment
and supersedes the old #cancel method.

If a payment for a Payment Intent was captured, it cannot be cancelled
anymore, but needs to be refunded.
For this refund we're using the default order cancel reason named after
`Spree::Payment::Cancellation::DEFAULT_REASON`.