Fix tests on Rails 7 because of selector not being found
Rails 7 made `#button_to` consistent [1], so that now it always render a
`<button>` HTML tag. Before that, its output was an `<input>` tag in
some circumstances.

Our tests were relying on the output to be an `<input>` tag, but we can
be generic and catch both the old and new behaviors.

See [2] for the line of code in solidus were the tag is generated.

[1] - https://github.com/rails/rails/pull/40747
[2] - https://github.com/solidusio/solidus/blob/350ed34e54fe27ff4e7ef8ea7478dbd5372796b9/backend/app/helpers/spree/admin/orders_helper.rb#L12