Remove the authorize action
Authorization in the current status requires direct action from the
user and the Stripe iframe, so it's not possible to acquire it from
the gateway and this method is not going to be used in this form for
now.

It will maybe re-introduced when reusable payment methods will be
added.