Make webhook endpoint specific to the payment method instance
The work on this PR allows registering different webhook endpoints for
each instance of Solidus Stripe payment methods. That's a scenario that
can help support marketplaces.

We're automatically generating a slug when a Stripe payment method is
created. Users must use that in the default
`/solidus_stripe/webhooks/:slug` endpoint to get the specific URL for
a payment method instance. The slug is stored on a separate
`SolidusStripe::WebhookEndpoint` model with a one-to-one relationship
with `SolidusStripe::PaymentMethod`. There's no single entry point to
create payment methods, so we must use an ActiveRecord callback to
persist the slug. We're using a slug and not the payment method id
to be resilient against database changes and to avoid guessing about
valid endpoint URLs.

As a consequence of bringing the webhook endpoint URL to the payment
method instance, we're also moving the configuration of the webhook
signing secret from a generic Rails credential to a payment method
preference.

We're now providing the slug to the
`SolidusStripe::Event.from_request` method to retrieve the associated
payment method and its signing secret. The payment method is now bundled
as part of the created Omnes event so that subscribers can act
differently in every case. For the test setup, the payment method is now
part of the `EventWithContextFactory` instances.

Closes #161