Add a partial for existing cards
Stripe has support for saving multiple types of payment methods.
Of the 20+ available we only implement "card" which is the most used
and can work as a reference for other types.

For types missing a partial a RuntimeError is raised with a message
that reports the name of the partial.

See https://stripe.com/docs/api/payment_methods/object#payment_method_object-type