Move the intent specific methods from PaymentMethod to the intent classes
Move the parallel methods for PaymentIntent and SetupIntent from
PaymentMethod to their respective classes.

That uncovered a hidden dependency on payment method for intents. The
newly added association to payment method makes it linear to have
multiple Stripe payment methods in the same store. Any potential bug
due to the previous ambiguity is solved.

With that fixed our custom "find or create" methods could be replaced
with vanilla ActiveRecord calls to `find_or_create_by`.

Co-Authored-By: Alberto Vena <kennyadsl@gmail.com>