Handle payment_intent.succeeded event
A `payment_intent.successful` event is published [1] when capturing a
payment intent from the Stripe dashboard. At that point, we need to
update the associated Solidus payment to be marked as completed.

However, the webhook could also be received after capturing the payment
from the Solidus admin interface, as it updates through Stripe's API.
Therefore, we need to do nothing if the payment is already completed.

The unit tests on the subscriber cover the handler's behavior, while at
the integration layer, we're making sure that it's actually reached when
the "payment_intent.succeeded" event is published. Besides, this PR has
been manually tested for different checkout flows.

Closes #179

[1] - https://stripe.com/docs/payments/payment-intents/verifying-status#webhooks