Fix `Gateway#purchase` not capturing the payment
On Solidus, `Gateway#purchase` is meant to authorize and capture a
payment in a single step. However, we were using that method only to
authorize (confirm, on Stripe) the payment.

We're now moving the authorize logic to a helper method shared by the
`#authorize` and `#purchase` methods, while we use the `automatic`
capture method on `#purchase`.

We're also unifying the format for the success messages recorded in the
response.

We're getting rid of the option to pass custom payment intent options on
`#authorize`. That was the only method in the gateway supporting them,
and we can add it later to all the methods if we feel that's needed.

Closes #232