Refactor void and credit methods in SolidusStripe::Gateway
This commit updates the void and credit methods in `SolidusStripe::Gateway` to
remove the `_source` parameter, which is no longer required when calling these
methods due to the `payment_profiles_supported?` being set to false.
This change is necessary because the `_source` parameter was previously defined
but not called when invoking these methods in Solidus.

Specifically, this change is made to align with the behavior of the:
 - void action invoked in Spree::Payment::Processing#void_transaction! (https://github.com/solidusio/solidus/blob/6d5580068d00c8fb711710a71e350e4c1358946c/core/app/models/spree/payment/processing.rb#L98)
 - credit action invoked in Spree::Refund#process! (https://github.com/solidusio/solidus/blob/6d5580068d00c8fb711710a71e350e4c1358946c/core/app/models/spree/refund.rb#L68)