Handle payment_intent.canceled event
A `payment_intent.canceled` event is published [1] when canceling a
payment intent from the Stripe dashboard. At that point, we need to
update the associated Solidus payment to be voided.

However, the webhook could also be received after voiding the payment
from the Solidus admin interface, as it updates through Stripe's API.
Therefore, we need to do nothing if the payment is already voided.

Closes #181

[1] - https://stripe.com/docs/api/events/types#event_types-payment_intent.canceled