Handle `charge.refunded` webhook event
Stripe sends a `charge.refunded` [1] event when a refund is
performed. When that happens, we need to accordingly update our data by
creating a `Spree::Refund` for the given payment.

Notice that there are two ways of generating Solidus refunds: directly
or through an RMA (Return Merchandise Authorization). A refund generated
from Stripe can only integrate with the first kind, as there's no
reliable way to hook into the second flow.

Stripe and Solidus support partial refunds. The event's `refund_amount`
field we receive contains the sum of all the performed refunds on the
payment, so we need to deduce the previously refunded amount before
generating the new one.

Stripe refunds are required to be associated with a `Spree::RefundReason`, so
we're seeding a record with a generic `Stripe refund` reason name. The
name to fetch can be configured through the `#default_refund_reason_name`
preference.

As noted in #262, we're not checking yet if an incoming refund has already
been processed in the past.

[1] - https://stripe.com/docs/api/events/types#event_types-charge.refunded

Closes #180