Re-introduce wallet logic to store `on_session` and `off_session` payment sources
After a successful payment confirmation through a payment intent with `setup_future_usage`
set to 'on_session' or 'off_session', the payment source (Stripe payment method)
can be considered valid for future reuse.

For more details, refer to the Stripe documentation:
https://stripe.com/docs/api/payment_intents/update#update_payment_intent-setup_future_usage

This commit reintroduces the functionality of adding the payment source to the user's
wallet after a successful payment, so that the user can reuse it for future
transactions without having to enter payment information again.