Add support for store credits in Stripe payment integration
This commit adds functionality to support the use of store credits when
processing payments through the Stripe payment method.
Specifically, the initialization of the Stripe elements and the creation of
payment intents have been updated to ensure that the correct `amount`
is charged to the user, taking into account any available store credits that have
been applied to the order from the `display_order_total_after_store_credit` method.