Avoid duplicated Solidus refunds
Stripe doesn't guarantee that a webhook event will be delivered only once,
so we need to be prepared to handle duplicated incoming refund events. Just
creating a Solidus copy whenever we receive a refund event is not enough.

On top of that, Stripe won't identify the refund that triggered the event. Both
the `charge.refunded` and `payment_intent.succeeded` (for partial captures,
which generates a refund for the remaining amount under the hood) webhooks only give
us information to retrieve the list of all refunds associated with a charge or
payment intent.

Because of this, we are now syncing all the refunds associated with a payment
intent at every webhook event. We keep track of the refunds already present on
Solidus by leveraging the `transaction_id` field on `Spree::Refund`, copying the
Stripe refund id as value.

On top of that, we need to account for refunds created from Solidus admin panel,
which calls the Stripe API. In this case, we need to avoid syncing the
refund back to Solidus on the subsequent webhook. We're marking those refunds
with a metadata field on Stripe, so we can exclude them from the sync.

Closes #262