Fix preference migration
The migration being changed was meant to update keys database records
for `Spree::Preference` records associated with Stripe payment methods
created using the `solidus_gateway` extension.

But in reality the preference keys would not be found by the `LIKE`
query in the migration. Keys looked like this:

    spree/payment_method/stripe_credit_card/server/1

Which would not be caught unless we used the SQL `%` operator at the end
of the query. See the commit body.