Allow the use of non-card Stripe payment methods
This update ensures that the Payment Intent created is compatible with
the created Stripe Payment Method from Stripe Elements by allowing for
identification and use of activated payment methods within the Stripe
dashboard.
This prevents any issues with compatibility when attaching a non-card
Stripe Payment Method with the new Stripe Payment Intent.
When retrieving or creating a Payment Intent, the payment_method_type [1]
of the Stripe payment method will always be included for compatibility
assurance.

Why we're not enabling the automatic_payment_methods [2]?
The automatic_payment_methods parameter is not being enabled due to
potential issues with existing payment sources in admin and in general
in case of using authorize/purchase from the Gateway.
Enabling this parameter would require a mandatory `return_url` during
confirmation, which poses problems on the admin-side where confirmation
and authorize/purchase are performed on the Gateway without the
possibility of specifying a `return_url`.
The case is different for the frontend-side, where the confirmation
and authorization occur on the client-side (and where we specify the
`return_url`).

The error message received in this scenario is:
"You must provide a return_url when confirming using automatic_payment_methods[enabled]=true."

In addition to addressing compatibility issues, this change will also
resolve #263 by enabling the possibility to specify supported payment
method types directly during the initialization of Stripe Elements.

[1] https://stripe.com/docs/api/payment_intents/object#payment_intent_object-payment_method_types
[2] https://stripe.com/docs/api/payment_intents/object#payment_intent_object-automatic_payment_methods-enabled